Avoid `setpgrp` if `setsid` fails
`setsid` throws [`EPERM`](https://man7.org/linux/man-pages/man2/setsid.2.html#ERRORS) if it's attempted for a process that's already a group leader. It is bound to fail if you run`setpgrp` before `setsid`. As a result, the `ruby-lsp` process goes into stopped state and the client waits forever for the booting to finish.