Add support for checking for pending migrations and running them (#503)
Add the ability to check for pending migrations and run them through the Rails runner client.

For checking if something is pending, we use the Rails API directly, which raises if there are any pending migrations.

For running them, we can't do it in process. The code that runs migrations is not designed to be invoked repeatedly and uses `load`, which means that we could end up with a memory bloat. I used `Open3.capture2` to run in a child process and grab the results.