Check for pending migrations when watcher is fired (#504)
Check for pending migrations, notify the user and offer to run them. The idea is that if any modifications are made to the `db/migrate` directory, we check if there are pending migrations.

If there are, we show a dialog and offers to run them. If the user decides to do it, we then fire a request to the server to run the migrations.

Note: we only show the dialog if a migration is deleted or created, otherwise it can become really annoying when you are editing the migration you just created.