Remove usage of present? from hover (#509)
We cannot use `present?` in the add-on side of the code because that method is defined by `activesupport` and that gem is not required on the LSP side.

It's only safe to use it in the runtime server side. This is currently breaking hover.

Note: tests won't catch this because in tests the Rails context is loaded through the dummy application, which makes tests pass despite this code breaking every time in real scenarios.