Unify all error handling for the server (#527)
Closes https://github.com/Shopify/ruby-lsp/issues/2886

We cannot use blanket rescue statements for the entire server because the behaviour for handling errors is different between requests and notifications.

For requests, we want to return `error: message` instead of the typicaly `result: something`. And for notifications, we cannot return anything, we can only log that an error occurred.

The approach that I took here is to create handlers that we can use both ourselves and in server add-ons, that provide appropriate error handling for requests or notifications.

I also created convenience  methods for returning responses or errors.