Check if global state is already initialized before creating listeners (#562)
We were originally assuming that the global state was always going to be present when a listener is created.

Due to the concurrent nature of the LSP, this is not always true. If the Rails + Tapioca add-on combination takes a while to finish booting, it's expected that the LSP may switch threads and a request will get executed before activating add-ons is complete.

I started returning early if the global state is not yet available.