Allow server to send asynchronous notifications to client (#559)
Previously, we were only allowing the server and its runtime add-ons to send simple strings to the stderr pipe, which would become log notifications.

To support initiating progress from the server, we will need to allow any type of LSP notification to be send asynchronously.

This PR changes our current logger thread to become a more general notifier thread, which will read any messages received from the server and push them to the outgoing queue.

To prevent creating a breaking change, I kept the `log_message` method, but changed it to build the right notification expected.