Start passing relevant capabilities to runtime server (#560)
To be able to support server initiated progress, it will need to know if the editor connecting to the LSP actually supports it, otherwise we risk sending notifications that the client cannot handle.

My proposal is that we start passing a JSON containing only the relevant client capabilities to the server, so that we can check what is enabled.