Handle render calls from controllerless template directories
In our app, we have an `app/views/shared/components` directory that
contains shared "component" partials. It doesn't have any matching
controller, and components can render other components, you just need
to specify the full path to the render call (we have a helper method for
that). We should make go to definition work for these cases as well.